requirements/common.txt
django>=1.5
django-cms>=3.0
django-sekizai